refactor(core): flatten `extensions/embed/` into `extensions/`
