+++ packages/core/src/extensions/embed-types.ts
